Rearrange files for future refactor.
