+++ .gitignore
@@ -56,3 +56,5 @@ venv.bak/
 
 # examples
 ABACUS/
+# temporary
+tests/
