+++ deepks/__init__.py
@@ -3,7 +3,7 @@ __author__ = "Yixiao Chen"
 try:
     from ._version import version as __version__
 except ImportError:
-    __version__ = 'unkown'
+    __version__ = 'unknown'
 
 __all__ = [
     "iterate",
